feat: refactor agent and veramo lifecycle
