+++ packages/dapp/src/app/[locale]/app/layout.tsx
@@ -10,10 +10,10 @@ export default async function AppLayout({
   children: React.ReactNode;
 }) {
   return (
-    <div className="flex h-full w-full flex-col px-4 sm:px-12">
+    <div className="flex h-full w-full flex-col px-4 sm:px-12 ">
       <AppNavbar />
       <div className="flex w-full flex-1 pb-14 pt-24 md:pb-0">
-        <div className="flex h-full flex-1 pb-4 pt-4 sm:pb-10 sm:pt-10 md:pb-16 md:pt-16 lg:pb-20 lg:pt-20">
+        <div className="flex h-full w-full flex-1 pb-4 pt-4 sm:pb-10 sm:pt-10 md:pb-16 md:pt-16 lg:pb-20 lg:pt-20">
           <CheckMetaMaskCompatibility />
           {children}
         </div>
