+++ packages/snap/src/utils/didUtils.ts
@@ -6,12 +6,8 @@ import type {
 import { BIP44CoinTypeNode } from '@metamask/key-tree';
 import { MetaMaskInpageProvider } from '@metamask/providers';
 import type { SnapsGlobalObject } from '@metamask/snaps-types';
-import type { IIdentifier } from '@veramo/core';
-import type { DIDResolutionResult } from 'did-resolver';
 
-import { getAgent } from '../veramo/setup';
-import { snapGetKeysFromAddress } from './keyPair';
-import { getCurrentNetwork } from './snapUtils';
+import VeramoService from '../veramo/Veramo.service';
 import { updateSnapState } from './stateUtils';
 
 export async function changeCurrentVCStore(params: {
@@ -26,67 +22,6 @@ export async function changeCurrentVCStore(params: {
   await updateSnapState(snap, state);
 }
 
-export async function getCurrentDidIdentifier(params: {
-  ethereum: MetaMaskInpageProvider;
-  state: MascaState;
-  snap: SnapsGlobalObject;
-  account: string;
-  bip44CoinTypeNode: BIP44CoinTypeNode;
-}): Promise<IIdentifier> {
-  const { ethereum, snap, state, account, bip44CoinTypeNode } = params;
-  const method = state.accountState[account].accountConfig.ssi.didMethod;
-  const agent = await getAgent(snap, ethereum);
-  switch (method) {
-    case 'did:pkh':
-    case 'did:ethr': {
-      const chainId = await getCurrentNetwork(ethereum);
-
-      if (method === 'did:pkh' && chainId !== '0x1' && chainId !== '0x89') {
-        throw new Error(
-          `Unsupported network with chainid ${chainId} for ${method}`
-        );
-      }
-      const identifier: IIdentifier = {
-        provider: method,
-        did:
-          method === 'did:ethr'
-            ? `did:ethr:${chainId}:${state.currentAccount}`
-            : `did:pkh:eip155:${chainId}:${state.currentAccount}`,
-        keys: [],
-        services: [],
-      };
-      return identifier;
-    }
-    case 'did:key:jwk_jcs-pub':
-    case 'did:key':
-    case 'did:jwk': {
-      const res = await snapGetKeysFromAddress({
-        bip44CoinTypeNode,
-        account,
-        snap,
-        state,
-      });
-
-      if (!res) throw new Error('Failed to get keys');
-
-      const identifier: IIdentifier = await agent.didManagerCreate({
-        provider: method === 'did:key:jwk_jcs-pub' ? 'did:key' : method,
-        kms: 'snap',
-        options: {
-          privateKeyHex: res.privateKey.slice(2),
-          keyType: method === 'did:key:jwk_jcs-pub' ? 'Secp256r1' : 'Secp256k1',
-          ...(method === 'did:key:jwk_jcs-pub' && { type: 'ebsi' }),
-        },
-      });
-
-      if (!identifier?.did) throw new Error('Failed to create identifier');
-      return identifier;
-    }
-    default:
-      throw new Error('Unsupported DID method');
-  }
-}
-
 export async function changeCurrentMethod(params: {
   snap: SnapsGlobalObject;
   ethereum: MetaMaskInpageProvider;
@@ -99,23 +34,6 @@ export async function changeCurrentMethod(params: {
     params;
   state.accountState[account].accountConfig.ssi.didMethod = didMethod;
   await updateSnapState(snap, state);
-  const identifier = await getCurrentDidIdentifier({
-    ethereum,
-    snap,
-    state,
-    account,
-    bip44CoinTypeNode,
-  });
+  const identifier = await VeramoService.getIdentifier();
   return identifier.did;
-}
-
-export async function resolveDid(params: {
-  did: string;
-  snap: SnapsGlobalObject;
-  ethereum: MetaMaskInpageProvider;
-}): Promise<DIDResolutionResult> {
-  const { did, snap, ethereum } = params;
-  const agent = await getAgent(snap, ethereum);
-  const didResolution = await agent.resolveDid({ didUrl: did });
-  return didResolution;
 }
